get coverage for default getEntry() method
