+++ api/all/src/test/java/io/opentelemetry/api/baggage/BaggageTest.java
@@ -9,6 +9,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.BiConsumer;
+import javax.annotation.Nullable;
 import org.junit.jupiter.api.Test;
 
 class BaggageTest {
@@ -29,13 +33,41 @@ class BaggageTest {
   }
 
   @Test
-  void getEntry() {
+  void getEntryDefault() {
     BaggageEntryMetadata metadata = BaggageEntryMetadata.create("flib");
-    try (Scope scope =
-        Context.root().with(Baggage.builder().put("a", "b", metadata).build()).makeCurrent()) {
-      Baggage result = Baggage.current();
-      assertThat(result.getEntry("a").getValue()).isEqualTo("b");
-      assertThat(result.getEntry("a").getMetadata().getValue()).isEqualTo("flib");
-    }
+    // Implementation that only implements asMap() which is used by getEntry()
+    Baggage baggage =
+        new Baggage() {
+
+          @Override
+          public Map<String, BaggageEntry> asMap() {
+            Map<String, BaggageEntry> result = new HashMap<>();
+            result.put("a", ImmutableEntry.create("b", metadata));
+            return result;
+          }
+
+          @Override
+          public int size() {
+            return 0;
+          }
+
+          @Override
+          public void forEach(BiConsumer<? super String, ? super BaggageEntry> consumer) {}
+
+          @Nullable
+          @Override
+          public String getEntryValue(String entryKey) {
+            return null;
+          }
+
+          @Override
+          public BaggageBuilder toBuilder() {
+            return null;
+          }
+        };
+
+    BaggageEntry entry = baggage.getEntry("a");
+    assertThat(entry.getValue()).isEqualTo("b");
+    assertThat(entry.getMetadata().getValue()).isEqualTo("flib");
   }
 }
